feat(web): show pass/fail % to tenths of a percent (#174)
This calculates the pertentage without rounding in the stats, and moves
the setting of how much to round it into the template.

I'm not sure why all the integer stats are in 32 bit instead of a more
standard usize or u64. I went with f32 for the ratio purely out of cargo
cult. My instinct would otherwise be to f64.