feat(web): add chart showing the content coverage
Show the (claimed) coverage of the nodes found by the census. Split the
keyspace into 256 1-byte-wide buckets. Then show how many nodes fully
cover that keyspace, and how much partial coverage there is of them.

This mostly reuses the existing data from the node radius chart, plus a
bit more precision in the radius calculation sent by the server.

In the tooltip: show the prefix, type, and value of the data that you
hover over.

Manually choose the ticks at the bottom of the x-axis, because I wanted
to show the extremes and one tick at the start of every nibble