chore: rename census table migration for correct ordering
The sea-orm tools will reorder based on date, which makes sense.
Oddly, the date of the 'create census table' migration comes before
other entities that it requires, like 'record'. So this change renames
the migration to an impossible date, but one that keeps it in the same
ordering that it was manually inserted at before the change.

An impossible date was chosen as a flag that we don't really know when
the migration occured.