feat: use specific sender ENR when noting txfr fail
Previously, audit_internal_failure tracked the node_id of the sending
node in each failure row. This left the precise information about the
sender ambiguous, because the ENR associated with a node evolves over
time. By pinning to a specific row in the 'record' table, we know the
exact ENR at the time of the failure. It also makes the query simpler,
when looking up the client that was sending failed data.

The schema migration also migrates the data, by replacing the node ID
with the most recent record ID. While technically inaccurate, it's
better than losing the data, and probably not worth doing more detailed
work to find the exact ENR at the time of the transfer for this
legacy data.