feat: in transfer failure records, link to ENR instead of node (#387)
Previously, audit_internal_failure tracked the node_id of the sending node in each failure row. This left the precise information about the sender ambiguous, because the ENR associated with a node evolves over time. By pinning to a specific row in the 'record' table, we know the exact ENR at the time of the failure. It also makes the SELECT queries simpler, when looking up the sending client type.

The schema migration also generates the record data, by replacing the node ID with the most recent record ID at the time of migration. While technically inaccurate, it's better than losing the data, and probably not worth doing more detailed work to find the exact ENR at the time of the transfer for this relatively short window of legacy data.

There is a new test to make sure that the record is stored properly.