feat(web): show full week in areaGraph weekly charts (#403)
Before this change, the graph would be zoomed in so that the oldest data
and newest data would define the extremes of the X-axis. Now, the x-axis
is stretched to the full week, no matter which data is available.

Note that there is no tooltip to display now, if you hover on the
right-hand side of the displayed data. So we end early when trying to
move the tooltip in that scenario.

Bonus:

* chore: have git ignore shell scripts
* docs: add areaGraph name to docstring