Name change of fluffy client to nimbus-portal (#419)
The fluffy client has been renamed to Nimbus Portal client.
This commit adapts glados to that name change.

The docker image is also updated with a new tag.

Replaced every usage of fluffy in the code with nimbus.
There appear to be several instances where the ENR is checked for
the "f" value in the client field. I kept that in place but also
added the "n" value. Currently 'f' is still written by Nimbus
Portal clients, but once this change gets deployed, we change
change it to write 'n' in the ENR. After that the check for 'f'
can be removed in glados.