perf(web): load radii of peers in audit details (#435)
This slightly changes behavior: it pulls the most recent census data
that is up to 15 minutes after the audit time.

This is faster because we can do the sorting & limiting of the census
nodes all in index by using the surveyed_at value directly, instead of
the computed absolute difference.

Note that although we LIMIT 1 inside the census_node lookup, that is
only limiting by record ID. Because each node may have multiple records,
we only want the latest record for each node. So we still have to order
by the time and apply DISTINCT node_id.